feat(DocumentDB Compatability): Renamed to django_documentdb; added NOTICE according to Apache 2.0
