+++ django_documentdb/base.py
