+++ django_documentdb/client.py
