+++ django_documentdb/creation.py
