+++ django_documentdb/dbapi.py
