+++ django_documentdb/fields/auto.py
