+++ django_documentdb/fields/duration.py
