+++ django_documentdb/introspection.py
