+++ django_documentdb/schema.py
