+++ django_documentdb/utils.py
